Added Heavy Crossbow and Tower Shield

+++ src/main/java/net/paweltl/pawelsweaponry/item/custom/CleaverItem.java
@@ -33,8 +33,8 @@ public class CleaverItem extends SwordItem {
         super(toolMaterial, attackDamage, attackSpeed, settings);
         this.attackDamage = (float)attackDamage + toolMaterial.getAttackDamage();
         ImmutableMultimap.Builder<EntityAttribute, EntityAttributeModifier> builder = ImmutableMultimap.builder();
-        builder.put(EntityAttributes.GENERIC_ATTACK_DAMAGE, new EntityAttributeModifier(ATTACK_DAMAGE_MODIFIER_ID, "Weapon modifier", (double)this.attackDamage, EntityAttributeModifier.Operation.ADDITION));
-        builder.put(EntityAttributes.GENERIC_ATTACK_SPEED, new EntityAttributeModifier(ATTACK_SPEED_MODIFIER_ID, "Weapon modifier", (double)attackSpeed, EntityAttributeModifier.Operation.ADDITION));
+        builder.put(EntityAttributes.GENERIC_ATTACK_DAMAGE, new EntityAttributeModifier(ATTACK_DAMAGE_MODIFIER_ID, "Weapon modifier", this.attackDamage, EntityAttributeModifier.Operation.ADDITION));
+        builder.put(EntityAttributes.GENERIC_ATTACK_SPEED, new EntityAttributeModifier(ATTACK_SPEED_MODIFIER_ID, "Weapon modifier", attackSpeed, EntityAttributeModifier.Operation.ADDITION));
         if (!CompatInit.isBetterCombatLoaded && CompatInit.isReachEntityAttributesLoaded) {
             builder.put(ReachEntityAttributes.REACH, new EntityAttributeModifier("Weapon modifier", -0.5, EntityAttributeModifier.Operation.ADDITION));
             builder.put(ReachEntityAttributes.ATTACK_RANGE, new EntityAttributeModifier("Weapon modifier", -0.5, EntityAttributeModifier.Operation.ADDITION));
